addressing two of the hundreds of warnings compiling on windows

+++ include/c74_min_atom.h
@@ -109,7 +109,7 @@ namespace min {
 		}
 		
 		operator int() const {
-			return atom_getlong(this);
+			return (int)atom_getlong(this);
 		}
 
 		operator long() const {
